VD/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.39.0
groq==0.11.0
httpx==0.27.0
python-multipart==0.0.12
python-dotenv==1.0.1
pydub==0.25.1
gtts==2.5.3
soundfile==0.12.1
numpy==1.26.4
requests==2.32.3
aiofiles==24.1.0